Added a Template for Signin/Registration Rules
I have added a template that gets included on both the signin screen, and the registration screen, that can define a list of rules, instructions and other info before a user gets to sign in or register. Place the templates into the Local folder on your site to customize your own content.