Modified dscan.py to work with latest streamer code
Modified the file to pull initial ReplayGain values on the first scan of the file. Because Scanning is now a little longer, if we can do this in the same pass it will save time. The code will still scan just the ReplayGain values as before if the value is set to Empty in the Admin panel.